Split AircraftMoveOrderTargeter to its own file
